Ops/ML Ops/saabmleval/requirements.txt
azure-functions==1.0.8
scikit-learn==0.20.3
numpy==1.16.4
gensim==3.7.1
pyodbc==4.0.30
spacy==2.1.8
nltk==3.4
azure.cognitiveservices.language.textanalytics==0.2.0
msrest==0.6.10
requests==2.22.0
pandas==0.24.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz#egg=en_core_web_sm==2.1.0
